Backend/Agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "RAG Agent for Book Content"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "openai-agents>=0.6.0",
    "litellm>=1.50.0",
    "cohere>=5.10.0",
    "qdrant-client>=1.12.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
]
